narrow type
