+++ src/redteam/constants.ts
@@ -1,6 +1,8 @@
-import { HARM_CATEGORIES } from './plugins/legacy/harmful';
+import { HARM_CATEGORIES } from './legacy/harmful';
 
 export const REDTEAM_MODEL = 'openai:chat:gpt-4o';
+export const LLAMA_GUARD_REPLICATE_PROVIDER =
+  'replicate:moderation:meta/meta-llama-guard-2-8b:b063023ee937f28e922982abdbf97b041ffe34ad3b35a53d33e1d74bb19b36c4';
 
 export const BASE_PLUGINS = [
   'contracts',
